frontend/api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "trackademic-api"
version = "1.0.0"
dependencies = [
    "azure-functions==1.18.0",
    "azure-functions-worker==1.3.0",
    "sqlalchemy==2.0.23",
    "python-jose[cryptography]==3.3.0",
    "cryptography==41.0.7",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "passlib[bcrypt]==1.7.4",
    "pyodbc==5.0.1"
]
